+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+### Added
+
+- Add JSON Schema evaluation against a reference. ([@skryukov])
+
 ### Fixed
 
 - Fix compatibility issues for Ruby < 3.1 and JRuby. ([@skryukov])